+++ app/src/test/kotlin/org/btcmap/db/table/event/SchemaTest.kt
@@ -19,6 +19,7 @@ class SchemaTest {
         assertEquals("website", WEBSITE)
         assertEquals("starts_at", STARTS_AT)
         assertEquals("ends_at", ENDS_AT)
+        assertEquals("updated_at", UPDATED_AT)
     }
 
     @Test
@@ -32,5 +33,6 @@ class SchemaTest {
         assert(CREATE.contains(WEBSITE))
         assert(CREATE.contains(STARTS_AT))
         assert(CREATE.contains(ENDS_AT))
+        assert(CREATE.contains(UPDATED_AT))
     }
 }